Update prerelease note needle for F40
The existing needle matches on bits of the keyboard layout
selector under the note and the Help button next to it, for some
reason (I don't recall why). The Help button is gone because we
killed installer help, and the keyboard layout indicator is not
always shown any more, there's no reason to match on it here
that I can recall.

Signed-off-by: Adam Williamson <awilliam@redhat.com>