desktop_background: handle KDE post-upgrade welcome center
When we test major KDE version upgrades, we get a post-upgrade
version of the welcome center again after the upgrade, which we
don't really expect. This breaks the background test because the
welcome center is in the way of the background, it doesn't seem
to break any other update tests. So let's just handle it here.
Fortunately this version has an "OK" button we can click.

Signed-off-by: Adam Williamson <awilliam@redhat.com>