ostree_build: force python3-dnf into built ostrees
This makes them diverge slightly from official ones, but...we
need it to run updvercheck.py, and doing that in a container is
a pain because we have to get the input files in. So, let's just
make sure it's in our ostrees.

It used to be in them anyway, but last night it stopped being in
Rawhide ones for some reason.

Signed-off-by: Adam Williamson <awilliam@redhat.com>