realmd_join_cockpit: do systemctl daemon-reload before starting
In https://openqa.fedoraproject.org/tests/2724005 , it seems that
cockpit.service was restarted by a systemctl daemon reload during
install of the necessary packages to enrol to the domain, which
causes the web UI to unexpectedly show a "Server has closed the
connection" error and fail the test.

It seems like we can mitigate this by triggering a daemon reload
manually before launching cockpit - the reload seems to be sort
of 'queued' at this point, I'm not sure why - so let's do that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>