Add a workaround needle for greyed-out lang select page
Since lorax-41.3-1.fc41, the anaconda language selection page
appears greyed out on Silverblue installer images for some reason
(it's fine on the Everything network install image we build).
Let's treat this as a soft failure for now - it still works, just
looks a bit weird - so we don't have failures on every update.

Signed-off-by: Adam Williamson <awilliam@redhat.com>