Tweak _boot_to_anaconda and g-i-s flow for new Rawhide g-i-s
We're dropping the live user mode patch from g-i-s downstream
because it's just too hard to maintain, apparently. So on Rawhide
the live image will boot to the welcome screen as normal, but
running the installer will give you newui rather than webui. If
you need a language other than English you have to sort it out
at the desktop before running the installer.

On first boot, g-i-s will show *all* screens, not skipping
language, keyboard layout or timezone, because we did not see
those in the installer.

This adapts the tests to handle the new flow, and should still
work with the other flows.

Signed-off-by: Adam Williamson <awilliam@redhat.com>