g-i-s: go back to relnum conditional, update comments
Changing to a conditional based on whether we saw webui broke the
ostree installer install tests, because they of course use the
same g-i-s even though they don't use webui. So, we have to go
back to the relnum-based conditional :/

Yes, this means we have redundant screens in g-i-s for install
paths that use gtkui to deploy GNOME, e.g. SB installer images,
but I can't see a good way to fix that. We need to show those
screens for the live install, which is the 'most important'
one, and we don't have an obvious way to show them for installs
that used webui but hide them for installs that used gtkui.

Signed-off-by: Adam Williamson <awilliam@redhat.com>