Create tests for remote desktop connections.
This adds a test for remote connection. It consists of a server part
in which Gnome RDP connection is set up, enabled, and the server is run,
and the client part that uses Connections as a client to connect to the
server.
The test is considered successful when the connection is established
and the user has logged into the system.

Fixes: https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/333