Fix KDE update test with KDE 6.2
The behavior after we click 'Restart to Install Updates Now'
button has changed in KDE 6.2. The default choice is no longer
'Install Updates and Restart' (even though that's what the button
says), and there is no timer. So if we don't click anything, the
confirm overlay just stays open forever.

This makes us click on the appropriate confirmation button if we
see it. We can also use this to make the test run a bit faster on
other releases, I guess.

Signed-off-by: Adam Williamson <awilliam@redhat.com>