Add several more tests to the desktop update test set
There's no obvious reason we're not also running these tests on
updates, so let's do it. We have to skip the advisory and UEFI
post checks for desktop_login as the last step of that test is
shutting down the system.

We leave out desktop_login for now because of
https://gitlab.gnome.org/GNOME/gjs/-/issues/647

Signed-off-by: Adam Williamson <awilliam@redhat.com>