Add FEDORA-2024-e331cd53ac as a workaround for F41
Another update depends on it. It's gone stable already, but we
are having issues with composes ATM so it hasn't made a compose.
It's safe to do this as we can be sure the depended-on update
will be in the next stable compose whenever it completes, so we
can't wind up in an inconsistent state.

Signed-off-by: Adam Williamson <awilliam@redhat.com>