Fix menu_launch_type for i3
Without this, all we did on i3 was hit alt-d and...do nothing,
we didn't type the app name.

I was confused at how we didn't notice this before, but it looks
like at present menu_launch_type isn't actually used in any
test we run on i3 (a lot of tests that use it to run a terminal
got a branch for i3 which uses alt-enter instead). This is an
obvious landmine, though, and it caused things to look weird
when rebasing #323 (which is how I noticed the bug).

This also refactors the sub to use the same logic for all
desktops, just with a different key for i3, since they really
all work the same. i3 doesn't need as much waiting as the other
desktops, probably, but it doesn't really hurt and keeps the code
simple.

Signed-off-by: Adam Williamson <awilliam@redhat.com>