Drop Cheese flatpak from F39 ostree installer builds
The others have all been rebased to F41, so we had to bump the
Platform version to 41, but since Cheese is kinda dead these
days, its flatpak hasn't been bumped, and that makes building
the F39 ostree installer image fail. To work around this, sed
it out of the pungi config.

Signed-off-by: Adam Williamson <awilliam@redhat.com>