select_disks: workaround failed first Destination click
In tests of the Rawhide anaconda update that ports to Wayland,
we often hit failures because the first attempt to click on
Installation Destination doesn't work. This only happened on prod
(not staging), it didn't happen every time (but quite often), and
we can't reproduce it manually, so it seems like a weird glitch
that we should just work around. Simply waiting a second and
clicking again seems to do the job, and should be safe even if
the first click works (the second click will just be on an empty
area of the Installation Destination screen, unless we have like
eight disks attached).

Signed-off-by: Adam Williamson <awilliam@redhat.com>