Rewrite tests from VNC to RDP
The VNC installation method has been replaced with a similar method
based on RDP. This PR uses the old mechanism and alters it for the
RDP method.

The test suite consists of two parts, a server and a client part.
The server is the real test on which Fedora Custom installation
is performed from the client via RDP.

The client is spinned off the pre-installed desktop image.

Fixes #345