Replace tour_start with a version with no OCR
OCR seems to be a common trigger of the mystery thread crash.
It's also not working properly anyway - the text we're trying
to match here is not in the current version of the tour but we
pass the needle match somehow. So, let's just ditch the attempt
to use OCR and go with regular area matches.

Signed-off-by: Adam Williamson <awilliam@redhat.com>