Retry live install tests 5 times
Unfortunately it doesn't look like qemu `-cpu host` really does
fully work around the grub2-mkconfig install hang issue, so we
need retries still :( As well as doing it for the update tests,
let's do it for the compose install tests. We use a substitution
to only increase the retry count for live images. We don't need
to set INSTALL_RETRY for other products - if it's not set then
RETRY winds up empty, and openQA treats that the same as it being
unset, which is fine.

Signed-off-by: Adam Williamson <awilliam@redhat.com>