select_cockpit_update: be more defensive against scrolls
This new way is working better, but a few times I've seen it fail,
I think because it's matching the 'updates' link while it's still
scrolling so when it clicks, the link has moved and the click
fails. Let's wait a bit after we see it, then match *again* and
click, instead of relying on the last match area.

Signed-off-by: Adam Williamson <awilliam@redhat.com>