Move testdata tarball into this repo
Again to reduce direct usage of pagure.io, move the testdata
tarball (used by download_testdata) directly into this repo. For
now it includes repository.tar.gz, which is silly, but I'll send
a follow-up PR to openqa_testdata to drop that file if this works
out. This includes a script to produce the tarball from the base
repository, and adjusts download_testdata to get the tarball from
autoinst_url and update its unpacking logic (as this version of
the tarball has a top-level directory).

Note the tarball in this commit is actually from the branch that
backs https://pagure.io/fedora-qa/openqa_testdata/pull-request/1
to reduce size.

Signed-off-by: Adam Williamson <awilliam@redhat.com>