startstop contacts: be consistent with contacts aaa_setup
@lruzicka recently made this test click through the setup page,
if it sees it...but he used new needles (which he didn't commit)
even though we already have those same needles as part of the
detailed contacts app test.

This setup page always appears, we don't need to check for it,
and we don't need separate needles to identify the page and to
click on the local address book option. All we need to do is the
same as we do in aaa_setup: unconditionally look for and click
the local address book option, using the already-existing needle.
So let's do that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>