blivet_btrfs_preserve_home: tweak root deletion
Trying to fix the frequent failures of this test, still. I don't
think we need the loop if we make sure to select the *parent*
entry in the list, which the needle tweaks should ensure, but
we might need to click twice to ensure it's selected and not
delete the entire btrfs volume by mistake, which is what we keep
doing.

Signed-off-by: Adam Williamson <awilliam@redhat.com>