Simplify RDP client flow and update needles
In two cases we don't need separate needles for identifying a
screen and then clicking something on it: we can just also use
the thing-to-click for identification purposes. Also remove the
connect_button-verify needle which has never matched (it matches
on the Verify button but has the tag for the Connect button, no
idea why), and update all the needles for the new GNOME fonts.

Signed-off-by: Adam Williamson <awilliam@redhat.com>