_boot_to_anaconda: drop installer launch retry loop
This seems to be sometimes *causing* problems now. In some cases,
the first launch actually worked but we don't wait long enough
for anaconda to show up, so we launch it twice, and that can
cause failures like https://bugzilla.redhat.com/2360859 .

I did a dry run for a few days on staging and just dropping the
loop entirely didn't seem to produce any failures-to-launch, so
maybe the various bugs we added this for in the first place have
all gone away? Let's try dropping it. If we run into failure-to
-launch problems again, we can add it back but bump the timeout
or something.

Signed-off-by: Adam Williamson <awilliam@redhat.com>