fifloader: introduce Flavors concept
I noticed that products with the same flavor often repeat the
same settings. Let's use our intermediate format to reduce this
repetition. We'll add a top-level dict for Flavors which allows
them to have settings, then in reverse_qol, for each product,
merge in the flavor's settings. Conflicts are resolved in favor
of the product.

Signed-off-by: Adam Williamson <awilliam@redhat.com>