Fix profile and group merging
Any of the four might not exist. At least *one* of the four has
to exist, but we can't rely on any particular one existing, so we
have to be very defensive. The only rule is that at least one
out of profiles and groups has to exist after merge.

Signed-off-by: Adam Williamson <awilliam@redhat.com>