fifloader: require group_name to be set via product
I've always hated the magic group name derivation in fifloader,
and it prevents it being used by anyone but Fedora. This fixes
that by requiring group_name to be set in the Products. This is
an executive decision, but it makes sense at least with how
Fedora uses openQA. If other users who might want to adopt
fifloader *don't* have a 100% assocation from products to job
groups, we can introduce more ways to set it in future, I guess.

I made it compulsory because it feels weird to have job templates
that aren't in groups...I don't even know how those get shown in
the web UI. But if there's a real use case for this we can make
it optional.

Signed-off-by: Adam Williamson <awilliam@redhat.com>