Make updatepkgs and updatepkgnames more precise for side tags
The way side tags work, every package in the distro is in the side
tag repo, so updatepkgs and updatepkgnames end up huge. To fix
this, we'll use a koji query for all packages that are actually in
the side tag to generate updatepkgnames.txt, and then use that to
filter the dnf command to generate updatepkgs.txt.

Signed-off-by: Adam Williamson <awilliam@redhat.com>