Add/update some needles for testing non-Server as Server
These are weird needles, but we do occasionally need to do this,
so I figured I'd add them. These are needed if you test an
Everything installer image as "Server-dvd-iso", which can be
useful to run all the former 'universal' tests on such an image.
Several will fail anyway, but a lot will run, if the appropriate
needles are in place.

Signed-off-by: Adam Williamson <awilliam@redhat.com>