Use Nehalem CPU for base_services_start on x86_64
Using cpu=host for this test on our new openQA workers makes it
fail, because mcelog.service doesn't want to run:

Jul 02 21:12:45 fedora mcelog[615]: mcelog: ERROR: AMD Processor family 25: mcelog does not support this processor.  Please use the edac_mce_amd module instead.

this seems like the simplest way to avoid that - let's just run
on a different CPU.

While I'm at it, let's do this more properly for apps_startstop.
It's wrong to hardcode Nehalem at the test level as we may want
to run it on other arches, so make it a variable defined at the
machine level. MINCPU is meant to be "minimum CPU" - a CPU with
baseline capabilities for the arch. For now it's defined as 'host'
on every other machine as I don't know if there is a baseline
qemu CPU model for aarch64 or Power, I'm asking those teams.

Using the same variable for base_services_start is really a bit
of a hack/workaround, but it seems silly to also define MCELOGCPU
or something with the exact same values...

Signed-off-by: Adam Williamson <awilliam@redhat.com>