Show details if an error appears.
When Software complains about something, it shows a grey banner
in the lower part with a button to click (Details).
The update test has not utilized this anyhow, so if it happened,
we would not know what the error was.
This fix makes sure to click on the Details button when it
appears, revealing the reason behind it, which could help
us understand what happened.