podman: run full system tests
Don't run only a minor subset of the tests instead run all of them. The
main advantage is that we get much better coverage and upstream don't
need to maintain an extra subset of tags. Of course tests will be much
slower now as it will now run like 11x the amount of test cases.

Fixes: https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/373

Signed-off-by: Paul Holzinger <pholzing@redhat.com>