_check_install_source: drop old stuff, update for dnf5
All the old patterns we accept and differing files they might be
in are obsolete now, I'm pretty sure - I checked that in F41
tests we get the most recent messages and they're all in syslog.
So drop all of those. But we need to add new handling for dnf5,
now anaconda switched to that and the messages changed again.

I can't find any messages in the dnf5 logs that confirm the
correct URL was used, unfortunately - these are the best I can
find.

Also, since 31cd5e7 , we won't get the base repository error
strings we're looking for any more, even if there *is* a problem
with the base repo. I don't think there's any single error we
can rely on getting any more, so we'll have to trust that the
other checks are sufficient, or figure out ones to add as we go
along if not.

Signed-off-by: Adam Williamson <awilliam@redhat.com>