RDP client: wait for server before pinging
This has always incorrectly been a race, it looks like, but for
some reason we were winning it before but we're losing it now.
The client seems to be pinging while the server's still typing
stuff into grub. So let's have the server set a mutex, and the
client wait for it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>