podman: invert bats' filter behavior, filter failing test
Until https://github.com/bats-core/bats-core/pull/1114 we can't
easily filter out specific tests in a bats run. So let's hack it
up with a sed that just negates the behavior of --filter - we
don't need to use its default behavior. Then we can use --filter
to filter out specific tests, like this. Use it to filter out
the test that's known to fail with recent kernels.

Signed-off-by: Adam Williamson <awilliam@redhat.com>