Work around another lock screen issue in desktop_login
We don't crash on the lock screen in Rawhide GNOME now, but my
fix isn't perfect - the pointer is somehow broken on the lock
screen, so we can't click anything. tab doesn't work either, so
we just can't switch user from the lock screen till this is
fixed better. So for GNOME on Rawhide, we'll skip switching from
the lock screen and just do another 'normal' switch.

Signed-off-by: Adam Williamson <awilliam@redhat.com>