Tweak _boot_to_anaconda web UI flow
We never get direct to installmethod any more, I'm pretty sure -
we always go through language selection. This may change if we
ever fix up the gnome-initial-setup stuff, but for now let's
simplify.

However, web UI now has a date/time page that shows on KDE but
apparently not on GNOME. So if we hit that, click through it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>