Add start-stop test for Kleopatra
This PR adds a start-stop test for Kleopatra that will replace
the KGpg application. It does not, however, remove the KGpg
start-stop test from the suite, as the KGpg is not gone yet.
We will remove it in a separate commit when time has come.

Fixes: https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/405

Change the start command to kleopatra

Move needle to generic space

Remove checking the selection box