Tweak CPU settings for ELN flavors
We used to use +QEMUCPU Haswell for BaseOS to override the old
machine default of Nehalem, because ELN has a higher CPU baseline
and doesn't run on Nehalem. We changed the machine default to
host, which is fine for ELN, so we should have dropped the
override. But there's a wrinkle - we set base_services_start to
run with MINCPU to fix mcelog.service failing with host, but if
we leave it that way the test will fail to run on ELN. So let's
override MINCPU to Haswell for ELN-on-x86_64 products. That ought
to be the right thing.

Signed-off-by: Adam Williamson <awilliam@redhat.com>